Use snapshot for insert-link tests
